Fixed tower placement bug
The centre of a tower is now placed where the mouse is clicked. Fixed the bug where the path and tower interference was not calculated properly. Made checking for interference between the path and circles more robust by writing a for loop that selects evenly-spaced points on the circle as the points to check (instead of hard coding specific points to check).